Panic when we can't convert between Prana and TiDB types.
This simplifies calling signatures quite a bit in numerous places. It is
a 100% mechanical refactoring.

The rationale here is that any type conversion failure will be our
error, not an end user error, so there's no benefit in returning it to
the user as they can't do anything about it. This is a common pattern in
Go. For example the "reflect" package returns no errors, requiring the
developer to correctly use the library and panicking on misuse.