Implement DESCRIBE statement
Adds a DESCRIBE statement that works for sources and materialized views.

The command takes a single value: the name of a source or MV in the
current scope.

It returns a row for every column with the column's name, type, and a flag
indicating if the column is a PK or not.