Fix StaticRows limit bug
(*StaticRows).GetLimit(limit int) ignores its limit parameter which can
cause the server to repeatedly send the same rows over and over again if
the requseted page size is smaller than the number of static rows.

This patch adds a cursor to SaticRows that tracks the progress of rows
sent.